Query fix, url fix
