+++ server/db/queries/userRepos/addRepoMembers.js
@@ -4,18 +4,18 @@ const { singleQuery } = require('../../baseQueries');
 
 const addRepoMembers = async ({ members }) => {
   members.map(async member => {
-    const values = [
-      uuidv4(),
-      member.githubId,
-      member.repoId,
-      member.userId || null,
-      member.userType,
-    ];
+    const values = [uuidv4(), member.githubId, member.repoId, member.userType];
 
     const queryText = `
       INSERT INTO
       user_repos(id, github_id, repo_id, user_id, user_type)
-      VALUES($1, $2, $3, $4, $5)
+      VALUES(
+        $1,
+        $2,
+        $3,
+        (SELECT u.id FROM users u WHERE u.github_id = $2),
+        $4
+      )
     `;
 
     await singleQuery({ queryText, values });
