+++ server/graphql/resolvers/users/deleteUser.js
@@ -44,7 +44,7 @@ const deleteUser = async (_, { authError, email, provider, userId }) => {
       username: '[deleted]',
     };
 
-    const { balance, githubId } = await getUserSettings({ userId });
+    const { balance } = await getUserSettings({ userId });
 
     if (balance > 0) throw new CustomError(remainingBalanceError);
 
@@ -52,7 +52,7 @@ const deleteUser = async (_, { authError, email, provider, userId }) => {
       await deleteCognitoUser({ email });
     }
 
-    await deleteRepoMembers({ githubId });
+    await deleteRepoMembers({ userId });
     await deleteUserLanguages({ userId });
     await deleteUserPullRequests({ userId });
     await transformUser({ data, userId });
