+++ server/helpers/index.js
@@ -1,17 +1,10 @@
-const {
-  arrayCheck,
-  CustomError,
-  errorLogger,
-  formatMemberList,
-  isUrl,
-} = require('./functions');
+const { arrayCheck, CustomError, errorLogger, isUrl } = require('./functions');
 const { sendEmail } = require('./emailHelper');
 
 module.exports = {
   arrayCheck,
   CustomError,
   errorLogger,
-  formatMemberList,
   isUrl,
   sendEmail,
 };
