+++ server/integrations/github/helpers.js
@@ -1,4 +1,5 @@
-const { CustomError } = require('../../helpers');
+const { CustomError, isUrl } = require('../../helpers');
+const { getRepoMembers } = require('./index');
 
 // ISSUE URL
 // Incoming:  https://github.com/tylermaran/cadl/issues/5
@@ -25,6 +26,24 @@ const formatIssueUrl = value => {
     };
   }
   throw new CustomError(`Not a valid issue url.`);
+};
+
+const formatMemberList = async ({ githubId, issueUrl, repoId, repoUrl }) => {
+  const { organization, repo } = issueUrl
+    ? formatIssueUrl(issueUrl)
+    : formatRepoUrl(repoUrl);
+  const githubMembers = await getRepoMembers({ organization, repo });
+  const formattedGithubMembers = githubMembers.map(({ id, type }) => ({
+    githubId: id,
+    repoId,
+    userType: type,
+  }));
+  const formattedRysolvOwner = {
+    githubId,
+    repoId,
+    userType: 'rysolv_owner',
+  };
+  return [...formattedGithubMembers, formattedRysolvOwner];
 };
 
 // PULL_REQUEST URL
@@ -57,7 +76,8 @@ const formatPullRequestUrl = value => {
 // REPO URL
 // Incoming:  https://github.com/NixOS or https://github.com/NixOS/nixpkgs-channels
 const formatRepoUrl = value => {
-  const { hostname, pathname } = new URL(value);
+  const newUrl = isUrl(value) ? value : `https://${value}`;
+  const { hostname, pathname } = new URL(newUrl);
 
   const containsGithub = hostname === 'github.com';
   const url = pathname.split('/');
@@ -86,6 +106,7 @@ const formatRepoUrl = value => {
 
 module.exports = {
   formatIssueUrl,
+  formatMemberList,
   formatPullRequestUrl,
   formatRepoUrl,
 };
